+++ app/ns-ui-widgets-category/dialogs/action-dialog/action-page.xml
@@ -1,4 +1,4 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd" loaded="onLoaded" navigatedTo="onNavigatedTo">
+<Page xmlns="http://www.nativescript.org/tns.xsd">
     <Page.actionBar>
 		<ActionBar title="Action"/>
 	</Page.actionBar>
