+++ app/ns-ui-widgets-category/dialogs/action-dialog/action-ts-page.ts
@@ -1,0 +1,24 @@
+// >> dialog-action-ts
+import { action, ActionOptions } from "tns-core-modules/ui/dialogs";
+
+export function showActionDialog() {
+    
+    
+    const actionOptions: ActionOptions = {
+        title: "Race selection",
+        message: "Choose your race",
+        cancelButtonText: "Cancel",
+        actions: ["Human", "Elf", "Dwarf", "Orc", "Unicorn"],
+        cancelable: true // Android only
+    };
+
+    action(actionOptions).then((result) => {
+        console.log("Dialog result: ", result);
+        if (result === "Options1") {
+            // Do action 1
+        } else if (result === "Option2") {
+            // Do action 2
+        }
+    });
+}
+// << dialog-action-ts
