+++ app/ns-ui-widgets-category/dialogs/action-dialog/action-ts-page.xml
