+++ app/ns-ui-widgets-category/dialogs/action-dialog/article.md
