+++ app/ns-ui-widgets-category/dialogs/alert-dialog/alert-page.js
@@ -1,26 +1,16 @@
-exports.onNavigatedTo = function (args) {
-    /*
-    // >> dialog-alert-web
-    alert("Your message").then(() => {
-        console.log("Dialog closed!");
-    });
-    // << dialog-alert-web
-    */
 
-    showAlertDialog();
-};
 
 function showAlertDialog() {
     // >> dialog-alert
     const alertOptions = {
-        title: "Your Title",
-        message: "Your message",
+        title: "Race selection",
+        message: "Race chosen: Unicorn",
         okButtonText: "OK",
         cancelable: false // [Android only] Gets or sets if the dialog can be canceled by taping outside of the dialog.
     };
 
     alert(alertOptions).then(() => {
-        console.log("Dialog closed!");
+        console.log("Race chosen!");
     });
     // << dialog-alert
 }
