+++ app/ns-ui-widgets-category/dialogs/alert-dialog/alert-page.xml
@@ -1,4 +1,4 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd" loaded="onLoaded" navigatedTo="onNavigatedTo">
+<Page xmlns="http://www.nativescript.org/tns.xsd" >
     <Page.actionBar>
 		<ActionBar title="Alert"/>
 	</Page.actionBar>
