+++ app/ns-ui-widgets-category/dialogs/alert-dialog/alert-ts-page.xml
