+++ app/ns-ui-widgets-category/dialogs/alert-dialog/article.md
@@ -1,0 +1,3 @@
+An Alert Dialog will notify the user for an action that has happened. It can be defined as:  
+<snippet id='dialog-alert'/>
+<snippet id='dialog-alert-ts'/>
