--- app/ns-ui-widgets-category/dialogs/alert/article.md
@@ -1,11 +1,0 @@
-An Alert Dialog will notify the user for an action that has happened.
-The `alert` method accepts string value to be displayed as the alert message or `AlertOptions` object 
-with keys `title`, `message`, `okBUttonText`, and `cancelable`(Android only property).
-
-Alert with string message (Web browser style).
-<snippet id='dialog-alert-web'/>
-<snippet id='dialog-alert-web-ts'/>
-
-Alert with `AlertOptions` object.
-<snippet id='dialog-alert'/>
-<snippet id='dialog-alert-ts'/>
