+++ app/ns-ui-widgets-category/dialogs/confirm-dialog/article.md
@@ -1,0 +1,3 @@
+A Confirm Dialog will expect the user to accept or reject the action that is about the happen. For example:
+<snippet id='dialog-confirm'/>
+<snippet id='dialog-confirm-ts'/>
