--- app/ns-ui-widgets-category/dialogs/confirm/article.md
@@ -1,9 +1,0 @@
-A Confirm Dialog will expect the user to accept or reject the action that is about the happen.
-
-Confirm with parameters (Web browser style).
-<snippet id='dialog-confirm-web'/>
-<snippet id='dialog-confirm-web-ts'/>
-
-Confirm with `ConfirmOptions` object.
-<snippet id='dialog-confirm'/>
-<snippet id='dialog-confirm-ts'/>
