+++ app/ns-ui-widgets-category/dialogs/login-dialog/article.md
@@ -1,0 +1,3 @@
+A Login Dialog will wait for the user to input their credentials: 
+<snippet id='dialog-login'/>
+<snippet id='dialog-login-ts'/>
