--- app/ns-ui-widgets-category/dialogs/login/article.md
@@ -1,9 +1,0 @@
-A Login Dialog will wait for the user to input their credentials.
-
-login with parameters (Web browser style).
-<snippet id='dialog-login-web'/>
-<snippet id='dialog-login-web-ts'/>
-
-login with `ConfirmOptions` object.
-<snippet id='dialog-login'/>
-<snippet id='dialog-login-ts'/>
