+++ app/ns-ui-widgets-category/dialogs/metadata.md
@@ -4,3 +4,4 @@ description: The Dialog module provides a functionality, which allows creating d
 position: 220
 slug: dialogs
 ---
+example-order: alert-dialog, action-dialog, confirm-dialog, login-dialog, prompt-dialog
