+++ app/ns-ui-widgets-category/dialogs/prompt-dialog/article.md
@@ -1,0 +1,3 @@
+A Prompt Dialog will request for an input. A basic definition might be: 
+<snippet id='dialog-prompt'/>
+<snippet id='dialog-prompt-ts'/>
