--- app/ns-ui-widgets-category/dialogs/prompt/article.md
@@ -1,9 +1,0 @@
-A Prompt Dialog will request for an input.
-
-Prompt with parameters (Web browser style).
-<snippet id='dialog-prompt-web'/>
-<snippet id='dialog-prompt-web-ts'/>
-
-Prompt with `ConfirmOptions` object.
-<snippet id='dialog-prompt'/>
-<snippet id='dialog-prompt-ts'/>
